App/requirements.txt
goose3==3.0.9
newspaper3k==0.2.6
nltk==3.2.5
pandas==0.22.0
Pyrebase==3.0.27
requests==2.11.1
tldextract==2.2.0
schedule==0.6.0
langdetect==1.0.7
humanize==0.5.1